Implemented budgets and improved dashboard layout

+++ BegrootSnelBackEnd/src/main/java/nl/hend/rm/dto/BudgetDto.java
@@ -4,4 +4,4 @@ import nl.hend.rm.entities.Category;
 
 import java.math.BigDecimal;
 
-public record BudgetDto(Category category, Long budgetId, BigDecimal amount, long year, long month) {}
+public record BudgetDto(Category category, Long budgetId, BigDecimal amount, int year, int month) {}
